dot-rewrite-engine/pyproject.toml
[project]
name = "dot-rewrite-engine"
version = "0.1.0"
description = "Study-state analysis engine for dot-rewrite notes"
requires-python = ">=3.12"
dependencies = [
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.0",
    "pydantic>=2.7",
    "numpy>=2.0",
    "scipy>=1.13",
    "scikit-learn>=1.5",
    "rank-bm25>=0.2.2",
    "gensim>=4.3",
    "nltk>=3.9",
    "python-igraph>=0.11",
    "leidenalg>=0.10",
    "networkx>=3.3",
    "spacy>=3.7",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.6"]

[project.scripts]
analyze-space = "engine.pipeline.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/engine"]

[tool.ruff]
line-length = 100
target-version = "py312"
